code duplication removed and IpcProvider fixed
