+++ packages/web3-providers/package.json
@@ -20,7 +20,6 @@
         "@types/node": "^10.12.18",
         "eventemitter3": "3.1.0",
         "lodash": "^4.17.11",
-        "oboe": "2.1.4",
         "url-parse": "1.4.4",
         "websocket": "git://github.com/frozeman/WebSocket-Node.git#browserifyCompatible",
         "xhr2-cookies": "1.1.0"
