+++ packages/web3-providers/src/providers/CustomProvider.js
@@ -70,16 +70,15 @@ export default class CustomProvider {
      *
      * @returns {Promise<any>}
      */
-    send(method, parameters) {
-        return this.sendPayload(JsonRpcMapper.toPayload(method, parameters)).then((response) => {
-            const validationResult = JsonRpcResponseValidator.validate(response);
+    async send(method, parameters) {
+        const response = this.sendPayload(JsonRpcMapper.toPayload(method, parameters));
+        const validationResult = JsonRpcResponseValidator.validate(response);
 
-            if (validationResult instanceof Error) {
-                throw validationResult;
-            }
+        if (validationResult instanceof Error) {
+            throw validationResult;
+        }
 
-            return response.result;
-        });
+        return response.result;
     }
 
     /**
