+++ packages/web3-providers/src/providers/MistEthereumProvider.js
@@ -11,15 +11,14 @@
     You should have received a copy of the GNU Lesser General Public License
     along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import JsonRpcMapper from '../mappers/JsonRpcMapper';
-import JsonRpcResponseValidator from '../validators/JsonRpcResponseValidator';
-import AbstractSocketProvider from '../../lib/providers/AbstractSocketProvider';
-
 /**
- * @file MistEthereumProvider
+ * @file MistEthereumProvider.js
  * @author Samuel Furter <samuel@ethereum.org>
  * @date 2019
  */
+
+import AbstractSocketProvider from '../../lib/providers/AbstractSocketProvider';
+
 export default class MistEthereumProvider extends AbstractSocketProvider {
     /**
      * @param {Web3EthereumProvider} connection
@@ -93,49 +92,6 @@ export default class MistEthereumProvider extends AbstractSocketProvider {
      */
     get connected() {
         return this.connection.isConnected();
-    }
-
-    /**
-     * Creates the JSON-RPC payload and sends it to the node.
-     *
-     * @method send
-     *
-     * @param {String} method
-     * @param {Array} parameters
-     *
-     * @returns {Promise<any>}
-     */
-    send(method, parameters) {
-        return this.sendPayload(JsonRpcMapper.toPayload(method, parameters)).then((response) => {
-            const validationResult = JsonRpcResponseValidator.validate(response);
-
-            if (validationResult instanceof Error) {
-                throw validationResult;
-            }
-
-            return response.result;
-        });
-    }
-
-    /**
-     * Creates the JSON-RPC batch payload and sends it to the node.
-     *
-     * @method sendBatch
-     *
-     * @param {AbstractMethod[]} methods
-     * @param {AbstractWeb3Module} moduleInstance
-     *
-     * @returns Promise<Object|Error>
-     */
-    sendBatch(methods, moduleInstance) {
-        let payload = [];
-
-        methods.forEach((method) => {
-            method.beforeExecution(moduleInstance);
-            payload.push(JsonRpcMapper.toPayload(method.rpcMethod, method.parameters));
-        });
-
-        return this.sendPayload(payload);
     }
 
     /**
